- Visão simples com CUDA, caso esteja instalado;
- Correção do display do FPS na interface;
- Melhorias gerais no código vision.cpp e .hpp;
- Muda as instruções do bash para sh no install_cuda.sh
- GMM desabilitado caso esteja usando cuda.

+++ cc/vision/visionGUI.hpp
@@ -12,7 +12,7 @@ class VisionGUI : public Gtk::VBox {
 		GMM *gmm;
 		Gtk::ToggleButton bt_HSV_calib;
 
-		int Img_id;
+		unsigned long Img_id;
 
 		bool HSV_calib_event_flag;
